profileImage validate done

+++ controllers/userControllers.js
@@ -101,9 +101,40 @@ module.exports.edituser = catchAsyncError(async (req, res, next) => {
     return next(new ErrorHandler("Please login to access the resource", 404));
   }
   if (req.files) {
+    const validMimeTypes = [
+      "image/jpeg",
+      "image/png",
+      "image/jpg",
+      "image/webp",
+    ];
+
+    if (!validMimeTypes.includes(req.files?.profileImage?.mimetype)) {
+      return next(
+        new ErrorHandler(
+          "Invalid file type. Only JPEG, PNG, JPG and WEBP files are allowed.",
+          500
+        )
+      );
+    }
+
+    const maxSize = 2 * 1024 * 1024; // 2MB
+
+    if (req.files?.profileImage?.size > maxSize) {
+      return next(
+        new ErrorHandler(
+          "File size exceeds the 2MB limit, Please select another file !",
+          500
+        )
+      );
+    }
+
     // old file delete code
-    if (user.profileImage.fileId !== "") {
-      await imagekit.deleteFile(user.profileImage.fileId);
+    try {
+      if (user.profileImage.fileId !== "") {
+        await imagekit.deleteFile(user.profileImage.fileId);
+      }
+    } catch (error) {
+      console.error("Failed to delete old profile image:", error);
     }
 
     const file = req.files.profileImage;
